Rename metriccards component.
